[nfc] Move implementation of IRUtils.h from IR.cpp to IRUtils.cpp
