+++ lib/IR/CMakeLists.txt
@@ -2,6 +2,7 @@
 add_library(IR
               IR.cpp
               IRGen.cpp
+              IRUtils.cpp
               IRBuilder.cpp
               Instrs.cpp
               GraphScheduler.cpp)
